+++ aldryn_sites/utils.py
@@ -51,13 +51,13 @@ def set_site_names(force=False):
                 site.save()
 
 
-def get_redirect_url(current_url, config, https=False):
+def get_redirect_url(current_url, config, https=None):
     """
     priorities are (primary domain and aliases are treated the same):
         exact redirect match > exact alias match > pattern redirect match > pattern alias match
     :param current_url: the url that is being called
     :param config: redirect configuration for this url
-    :param want_https: whether redircts should go to https
+    :param want_https: whether redirects should go to https (None keeps the current scheme)
     :return: None for no redirect or an url to redirect to
     """
     primary_domain = config['domain']
@@ -66,7 +66,10 @@ def get_redirect_url(current_url, config, https=False):
     redirect_domains = set(config.get('redirects', []))
     redirect_domain_patterns = compile_regexes(redirect_domains)
     url = yurl.URL(current_url)
-    target_scheme = 'https' if https else 'http'
+    if https is None:
+        target_scheme = url.scheme
+    else:
+        target_scheme = 'https' if https else 'http'
     redirect_url = None
     if url.is_host_ip() or url.is_host_ipv4():
         # don't redirect for ips
@@ -75,7 +78,7 @@ def get_redirect_url(current_url, config, https=False):
         # exact host and scheme match: Nothing to do
         return
     if url.host in domains and url.scheme != target_scheme:
-        # exact alias match, but scheme mismatch: redirect to change scheme
+        # exact alias match, but scheme mismatch: redirect to changed scheme
         redirect_url = url.replace(scheme=target_scheme)
     elif url.host in redirect_domains:
         # exact redirect match: redirect